travel-backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
Pillow
python-dotenv